+++ Content.Server/_RF/NPC/Queries/Filters/ToolQualityFilter.cs
@@ -1,13 +1,15 @@
 using System.Linq;
 using Content.Server.NPC;
+using Content.Shared.Tools;
 using Content.Shared.Tools.Components;
+using Robust.Shared.Prototypes;
 
 namespace Content.Server._RF.NPC.Queries.Filters;
 
 /// <summary>
 /// Filters tools with a certain quality
 /// </summary>
-public sealed partial class ToolQualityKeyFilter : RfUtilityQueryFilter
+public sealed partial class ToolQualityFilter : RfUtilityQueryFilter
 {
     private EntityQuery<ToolComponent> _toolQuery;
 
@@ -15,9 +17,7 @@ public sealed partial class ToolQualityKeyFilter : RfUtilityQueryFilter
     /// Quality to be filtered out
     /// </summary>
     [DataField(required: true)]
-    public string QualityKey = default!;
-
-    private string? _quality;
+    public ProtoId<ToolQualityPrototype> Quality;
 
     public override void Initialize(IEntityManager entManager)
     {
@@ -25,14 +25,8 @@ public sealed partial class ToolQualityKeyFilter : RfUtilityQueryFilter
         _toolQuery = entManager.GetEntityQuery<ToolComponent>();
     }
 
-    public override bool Startup(NPCBlackboard blackboard)
-    {
-        return blackboard.TryGetValue(QualityKey, out _quality, EntityManager);
-    }
-
     public override bool Filter(EntityUid uid, NPCBlackboard blackboard)
     {
-        return _toolQuery.TryComp(uid, out var comp)
-               && comp.Qualities.ToList().Contains(_quality!);
+        return _toolQuery.TryComp(uid, out var comp) && comp.Qualities.ToList().Contains(Quality);
     }
 }
