+++ Resources/Prototypes/_RF/NPCs/utility_queries.yml
@@ -12,69 +12,6 @@
       curve: !type:PresetCurve
         preset: TargetDistance
 
-- type: utilityQuery
-  id: TargetMaterials
-  query:
-    - !type:MapComponentQuery
-      components:
-        - type: Stack
-          stackType: Steel # Just serialization stuff
-    - !type:MaterialKeyFilter
-      targetKey: Material
-      amountKey: Amount
-      invert: true
-    - !type:InventoryFilter
-    - !type:InHandsFilter
-  considerations:
-    - !type:NormTargetDistanceCon
-      curve: !type:PresetCurve
-        preset: TargetDistance
-
-- type: utilityQuery
-  id: TargetTags
-  query:
-    - !type:MapComponentQuery
-      components:
-        - type: Tag
-    - !type:TagKeyFilter
-      tagsKey: Tags
-      requireAllKey: AllTags
-      invert: true
-    - !type:InventoryFilter
-    - !type:InHandsFilter
-  considerations:
-    - !type:NormTargetDistanceCon
-      curve: !type:PresetCurve
-        preset: TargetDistance
-
-- type: utilityQuery
-  id: TargetTool
-  query:
-    - !type:MapComponentQuery
-      components:
-        - type: Tool
-    - !type:ToolQualityKeyFilter
-      qualityKey: ToolQuality
-      invert: true
-    - !type:InventoryFilter
-    - !type:InHandsFilter
-  considerations:
-    - !type:NormTargetDistanceCon
-      curve: !type:PresetCurve
-        preset: TargetDistance
-
-- type: utilityQuery
-  id: TargetComponent
-  query:
-    - !type:ComponentKeyQuery
-      key: Component
-    - !type:InventoryFilter
-    - !type:InHandsFilter
-  considerations:
-    - !type:NormTargetDistanceCon
-      curve: !type:PresetCurve
-        preset: TargetDistance
-
 # Attack task
 
 - type: utilityQuery
